Update test_course.py, json.py
Fix typos and doc and remove print statements from json.py
move BaseCourseViewTestCase to test_cases.py and resolve redundant code
